// src/types/clinicalInfo.ts
export interface ClinicalInfo {
  id: string;
  insertion_date?: string;
  appliance_fit?: string;
  design_feedback?: string;
  occlusion?: string;
  esthetics?: string;
  adjustments_made?: string;
  material?: string;
  shade?: string;
  report_card_id?: string;
  created_at: string;
  updated_at: string;
}