// src/types/designInfo.ts
export interface DesignInfo {
  id: string;
  design_date: string;
  appliance_type?: string;
  upper_treatment?: string;
  lower_treatment?: string;
  upper_design_name?: string;
  lower_design_name?: string;
  screw?: string;
  implant_library?: string;
  teeth_library?: string;
  actions_taken?: string;
  report_card_id?: string;
  created_at: string;
  updated_at: string;
}